requirements.txt
smartapi-python==1.3.4
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
pyotp==2.9.0
websocket-client==1.7.0
requests==2.31.0
logzero==1.7.0
schedule==1.2.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
